Move WebSocketMessageUtils
